+++ .gitignore
@@ -1,0 +1,3 @@
+__pycache__/*
+detectors/__pycache__/*
+particles/__pycache__/*
